requirements.txt
bibtexparser==1.1.0
future==0.18.2
pyparsing==2.4.7
python-frontmatter==0.5.0
PyYAML==5.1
six==1.15.0
